refactor(wrapped-keys-lit-actions): LIT-3959 - Restructure directory layout to make file purposes explicit from their location in the package
