+++ packages/wrapped-keys-lit-actions/src/lib/internal/ethereum/signTransaction.js
