+++ packages/wrapped-keys-lit-actions/src/lib/self-executing-actions/common/exportPrivateKey.js
@@ -1,22 +1,20 @@
-const { signMessageSolanaKey } = require('./internal/signMessage');
 const {
   getDecryptedKeyToSingleNode,
-} = require('../common/internal/getDecryptedKeyToSingleNode');
-const { removeSaltFromDecryptedKey } = require('../utils');
+} = require('../../internal/common/getDecryptedKeyToSingleNode');
+const { removeSaltFromDecryptedKey } = require('../../utils');
 
-/* global accessControlConditions, ciphertext, dataToEncryptHash, messageToSign, Lit */
+/* global accessControlConditions, ciphertext, dataToEncryptHash, Lit */
 
 /**
  *
- * Bundles solana/web3.js package as it's required to sign a message with the Solana wallet which is also decrypted inside the Lit Action.
+ * Exports the private key after decrypting and removing the salt from it.
  *
  * @jsParam pkpAddress - The Eth address of the PKP which is associated with the Wrapped Key
  * @jsParam ciphertext - For the encrypted Wrapped Key
  * @jsParam dataToEncryptHash - For the encrypted Wrapped Key
- * @jsParam messageToSign - The unsigned message to be signed by the Wrapped Key
  * @jsParam accessControlConditions - The access control condition that allows only the pkpAddress to decrypt the Wrapped Key
  *
- * @returns { Promise<string> } - Returns a message signed by the Solana Wrapped key. Or returns errors if any.
+ * @returns { Promise<string> } - Returns a decrypted private key.
  */
 
 (async () => {
@@ -33,13 +31,7 @@ const { removeSaltFromDecryptedKey } = require('../utils');
     }
 
     const privateKey = removeSaltFromDecryptedKey(decryptedPrivateKey);
-
-    const signature = await signMessageSolanaKey({
-      messageToSign,
-      privateKey,
-    });
-
-    Lit.Actions.setResponse({ response: signature });
+    Lit.Actions.setResponse({ response: privateKey });
   } catch (err) {
     Lit.Actions.setResponse({ response: `Error: ${err.message}` });
   }
