+++ packages/wrapped-keys-lit-actions/src/lib/self-executing-actions/ethereum/signMessageWithEncryptedEthereumKey.js
@@ -1,20 +1,23 @@
 const {
   getDecryptedKeyToSingleNode,
-} = require('./internal/getDecryptedKeyToSingleNode');
-const { removeSaltFromDecryptedKey } = require('../utils');
+} = require('../../internal/common/getDecryptedKeyToSingleNode');
+const {
+  signMessageEthereumKey,
+} = require('../../internal/ethereum/signMessage');
+const { removeSaltFromDecryptedKey } = require('../../utils');
 
-/* global accessControlConditions, ciphertext, dataToEncryptHash, Lit */
+/* global accessControlConditions, ciphertext, dataToEncryptHash, messageToSign, Lit */
 
 /**
- *
- * Exports the private key after decrypting and removing the salt from it.
+ * Signs a message with the Ethers wallet which is also decrypted inside the Lit Action.
  *
  * @jsParam pkpAddress - The Eth address of the PKP which is associated with the Wrapped Key
  * @jsParam ciphertext - For the encrypted Wrapped Key
  * @jsParam dataToEncryptHash - For the encrypted Wrapped Key
+ * @jsParam messageToSign - The unsigned message to be signed by the Wrapped Key
  * @jsParam accessControlConditions - The access control condition that allows only the pkpAddress to decrypt the Wrapped Key
  *
- * @returns { Promise<string> } - Returns a decrypted private key.
+ * @returns { Promise<string> } - Returns a message signed by the Ethers Wrapped key. Or returns errors if any.
  */
 
 (async () => {
@@ -31,7 +34,13 @@ const { removeSaltFromDecryptedKey } = require('../utils');
     }
 
     const privateKey = removeSaltFromDecryptedKey(decryptedPrivateKey);
-    Lit.Actions.setResponse({ response: privateKey });
+
+    const signature = await signMessageEthereumKey({
+      privateKey,
+      messageToSign,
+    });
+
+    Lit.Actions.setResponse({ response: signature });
   } catch (err) {
     Lit.Actions.setResponse({ response: `Error: ${err.message}` });
   }
