+++ packages/wrapped-keys-lit-actions/src/lib/self-executing-actions/solana/signMessageWithEncryptedSolanaKey.js
@@ -1,13 +1,14 @@
-const { signMessageEthereumKey } = require('./internal/signMessage');
 const {
   getDecryptedKeyToSingleNode,
-} = require('../common/internal/getDecryptedKeyToSingleNode');
-const { removeSaltFromDecryptedKey } = require('../utils');
+} = require('../../internal/common/getDecryptedKeyToSingleNode');
+const { signMessageSolanaKey } = require('../../internal/solana/signMessage');
+const { removeSaltFromDecryptedKey } = require('../../utils');
 
 /* global accessControlConditions, ciphertext, dataToEncryptHash, messageToSign, Lit */
 
 /**
- * Signs a message with the Ethers wallet which is also decrypted inside the Lit Action.
+ *
+ * Bundles solana/web3.js package as it's required to sign a message with the Solana wallet which is also decrypted inside the Lit Action.
  *
  * @jsParam pkpAddress - The Eth address of the PKP which is associated with the Wrapped Key
  * @jsParam ciphertext - For the encrypted Wrapped Key
@@ -15,7 +16,7 @@ const { removeSaltFromDecryptedKey } = require('../utils');
  * @jsParam messageToSign - The unsigned message to be signed by the Wrapped Key
  * @jsParam accessControlConditions - The access control condition that allows only the pkpAddress to decrypt the Wrapped Key
  *
- * @returns { Promise<string> } - Returns a message signed by the Ethers Wrapped key. Or returns errors if any.
+ * @returns { Promise<string> } - Returns a message signed by the Solana Wrapped key. Or returns errors if any.
  */
 
 (async () => {
@@ -33,9 +34,9 @@ const { removeSaltFromDecryptedKey } = require('../utils');
 
     const privateKey = removeSaltFromDecryptedKey(decryptedPrivateKey);
 
-    const signature = await signMessageEthereumKey({
-      privateKey,
+    const signature = await signMessageSolanaKey({
       messageToSign,
+      privateKey,
     });
 
     Lit.Actions.setResponse({ response: signature });
